Fixed Jellyfish world freeze

+++ src/main/java/com/minecraftabnormals/upgrade_aquatic/common/entities/jellyfish/ai/BoxJellyfishHuntGoal.java
@@ -12,12 +12,12 @@ import java.util.EnumSet;
 public class BoxJellyfishHuntGoal extends Goal {
 	private final BoxJellyfishEntity hunter;
 	private int noSightTicks;
-	
+
 	public BoxJellyfishHuntGoal(BoxJellyfishEntity hunter) {
 		this.hunter = hunter;
 		this.setMutexFlags(EnumSet.of(Flag.MOVE, Flag.LOOK));
 	}
-	
+
 	@Override
 	public boolean shouldExecute() {
 		LivingEntity target = this.hunter.getAttackTarget();
@@ -28,12 +28,12 @@ public class BoxJellyfishHuntGoal extends Goal {
 		}
 		return false;
 	}
-	
+
 	@Override
 	public boolean shouldContinueExecuting() {
 		return super.shouldContinueExecuting() && this.noSightTicks < 20;
 	}
-	
+
 	@Override
 	public void startExecuting() {
 		this.hunter.setHuntingCooldown();
@@ -44,20 +44,20 @@ public class BoxJellyfishHuntGoal extends Goal {
 	public void tick() {
 		LivingEntity target = this.hunter.getAttackTarget();
 		Vector3d distance = new Vector3d(target.getPosX() - this.hunter.getPosX(), target.getPosY() - this.hunter.getPosY(), target.getPosZ() - this.hunter.getPosZ());
-		
-		float pitch = -((float) (MathHelper.atan2(distance.getY(), (double) MathHelper.sqrt(distance.getX() * distance.getX() + distance.getZ() * distance.getZ())) * (double) (180F / (float) Math.PI)));
+
+		float pitch = -((float) (MathHelper.atan2(distance.getY(), MathHelper.sqrt(distance.getX() * distance.getX() + distance.getZ() * distance.getZ())) * (double) (180F / (float) Math.PI)));
 		float yaw = (float) (MathHelper.atan2(distance.getZ(), distance.getX()) * (double) (180F / (float) Math.PI)) - 90.0F;
-		
+
 		this.hunter.lockedRotations[0] = MathHelper.wrapDegrees(yaw);
 		this.hunter.lockedRotations[1] = MathHelper.wrapDegrees(pitch + 90.0F);
-		
+
 		float[] rotations = this.hunter.getRotationController().getRotations(1.0F);
-		
-		if(!this.hunter.canEntityBeSeen(target)) {
+
+		if (!this.hunter.canEntityBeSeen(target)) {
 			this.noSightTicks++;
 		}
-		
-		if(this.hunter.isNoEndimationPlaying() && Math.abs(rotations[0] - this.hunter.lockedRotations[0]) < 7.5F) {
+
+		if (this.hunter.isNoEndimationPlaying() && Math.abs(rotations[0] - this.hunter.lockedRotations[0]) < 7.5F) {
 			NetworkUtil.setPlayingAnimationMessage(this.hunter, BoxJellyfishEntity.SWIM_ANIMATION);
 		}
 	}
